fix(onchain): dl1 commit-index route + point e2e sync polls at it
The e2e DL1-sync gates (waitForDl1Sync / waitForDl1AssetSync) polled
each DL1's /v1/onchain fiberCommits/assetCommits — a cumulative
surface that OnChain v2 turned into a vanishing per-batch delta, so
every sync gate timed out (all e2e lanes red).

- extract the fold/heal cache from Validator into a shared
  CommitIndexService; the DL1 wires ONE instance into both
  validateUpdate and a new GET /v1/commit-index route, so the
  ingestion gate and the polled sync surface observe the same folded
  view (and polling the route drives the same refresh)
- harness: sync helpers now poll /data-application/v1/commit-index
  (typed locally until the SDK ships CommitIndexResponse)
- fix heal client URL: custom routes mount under /data-application —
  the previous path 404'd every heal
- DL1 OpenAPI contract gains the commit-index endpoint (4 endpoints);
  docs regenerated

sharedData 651/651, currencyL0 8/8, e2e-test tsc clean.
